+++ pages/apps-web/esbuild.dist.mjs
@@ -1,15 +1,17 @@
 import esbuild from 'esbuild';
-import fs from 'fs';
-import path from 'path';
+import { copy } from 'esbuild-plugin-copy';
+import { wasmLoader } from 'esbuild-plugin-wasm';
+import { readFileSync, writeFileSync } from 'fs';
+import { resolve, dirname } from 'path';
 import { fileURLToPath } from 'url';
 import { htmlPlugin } from '@craftamap/esbuild-plugin-html';
 import { parse } from 'node-html-parser';
 
 const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const __dirname = dirname(__filename);
 
-const indexTemplate = fs.readFileSync(
-  path.resolve(__dirname, 'public/index.html'),
+const indexTemplate = readFileSync(
+  resolve(__dirname, 'public/index.html'),
   'utf8',
 );
 
@@ -20,32 +22,50 @@ indexDOM.getElementsByTagName('script').forEach((script) => {
   }
 });
 
+const plugins = [
+  copy({
+    resolveFrom: 'cwd',
+    assets: [
+      {
+        from: ['./public/assets/**/*'],
+        to: ['./dist/assets'],
+      },
+    ],
+  }),
+  htmlPlugin({
+    files: [
+      {
+        entryPoints: ['src/index.ts'],
+        filename: 'index.html',
+        scriptLoading: 'module',
+        htmlTemplate: indexDOM.toString(),
+      },
+    ],
+  }),
+  wasmLoader({ mode: 'deferred' }),
+];
+
 const common = {
   preserveSymlinks: true,
   treeShaking: true,
   metafile: true,
   minify: true,
+  metafile: true,
   bundle: true,
   format: 'esm',
   platform: 'browser',
   target: 'esnext',
 };
 
-esbuild.build({
-  ...common,
-  entryPoints: ['src/index.ts'],
-  entryNames: 'bundle-[hash]',
-  outdir: 'dist/',
-  plugins: [
-    htmlPlugin({
-      files: [
-        {
-          entryPoints: ['src/index.ts'],
-          filename: 'index.html',
-          scriptLoading: 'module',
-          htmlTemplate: indexDOM.toString(),
-        },
-      ],
-    }),
-  ],
-});
+esbuild
+  .build({
+    ...common,
+    entryPoints: ['src/index.ts'],
+    entryNames: 'bundle-[hash]',
+    outdir: 'dist/',
+    plugins: plugins,
+  })
+  .then(({ metafile }) => {
+    writeFileSync('build-meta.json', JSON.stringify(metafile));
+  })
+  .catch(() => process.exit(1));